+++ MoltenVK/MoltenVK/Utility/MVKEnvironment.h
@@ -248,9 +248,9 @@ void mvkSetMVKConfiguration(MVKConfiguration* pMVKConfig);
 #   define MVK_CONFIG_TEXTURE_1D_AS_2D    1
 #endif
 
-/** Preallocate descriptors when creating VkDescriptorPool. Disabled by default. */
+/** Preallocate descriptors when creating VkDescriptorPool. Enabled by default. */
 #ifndef MVK_CONFIG_PREALLOCATE_DESCRIPTORS
-#   define MVK_CONFIG_PREALLOCATE_DESCRIPTORS    0
+#   define MVK_CONFIG_PREALLOCATE_DESCRIPTORS    1
 #endif
 
 /** Use pooling for command resources in a VkCommandPool. Enabled by default. */